Node App
We now have a node js application that creates a local http server
running on port 8888. It displays and renders the html css, and js
associated with the webpage, and supports full navigation on that local
server. We need to hookup MongoDB and create a document creator, as well
as a table that shows all the available user in the system.